Add ability to set the maxUnavailable for the prometheus-collector daemonset
On large cluster with many nodes updating the prometheus-collector-node
pods one at a time can take a long time and will timeout frequently.

This change allows for the maxUnavailable to be set to make the deployment
of the daemonset faster and not timeout.  The default was also set to be
100% so all the nodes will be updated at once.