Test: Replace SP with managed identity, add test retries, increase wait time for windows containers (#866)
[comment]: # (Note that your PR title should follow the conventional
commit format: https://conventionalcommits.org/en/v1.0.0/#summary)
# PR Description
- Replace AMW querying with managed identity instead of a service
principal
- Add FlakeAttempts() to certain tests to enable retries
- Increase wait time for windows containers to fully start up before
running tests
- Add targetallocator watcher restart error to be ignored in the
container logs tests as this is expected
- Update go.mod/go.sum to latest packages for vulnerabilities

[comment]: # (The below checklist is for code changes. Not all boxes
necessarily need to be checked. Build, doc, and template changes do not
need to fill out the checklist.)
# Tests Checklist

- [x] Have end-to-end Ginkgo tests been run on your cluster and passed?
To bootstrap your cluster to run the tests, follow [these
instructions](/otelcollector/test/README.md#bootstrap-a-dev-cluster-to-run-ginkgo-tests).
  - Labels used when running the tests on your cluster:
    - [x] `operator`
    - [x] `windows`
    - [x] `arm64`
    - [ ] `arc-extension`
- [x] Have new tests been added? For features, have tests been added for
this feature? For fixes, is there a test that could have caught this
issue and could validate that the fix works?