fix: very high CPU issue with openmetrics exposition (#1162)
Change the `EnableCreatedTimestampZeroIngestion` setting in the
prometheus scrape manager to use the otelcollector feature flag that has
the same name and is disabled by default.

This had introduced new parsing during a scrape loop that resulted in
very high CPU at higher metric volumes.

Reverts this PR:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36660
so that the functionality for this is the same as before v0.121.0``