feat: add support for Container Storage's storage-operator
This commit integrates the control plane and data plane metrics
support for storage-operator, which meant to be a v2 for ACStor.
Data Plane metrics:
    disk_read_operations_completed_total
    disk_write_operations_completed_total
    disk_read_operations_time_seconds_total
    disk_write_operations_time_seconds_total
    disk_read_bytes_total
    disk_written_bytes_total
    disk_reads_merged_total
    disk_writes_merged_total
    disk_io_now
    disk_io_time_seconds_total
    disk_io_time_weighted_seconds_total
    disk_discard_operations_completed_total
    disk_discards_merged_total
    disk_discarded_sectors_total
    disk_discard_operations_time_seconds_total
    disk_flush_requests_total
    disk_flush_requests_time_seconds_total
These metrics are per user volume created (PV), thus the number
of volumes whill determine the cardinality of these metrics

Control Plane metrics:
    rpc_server_duration_milliseconds_bucket

    This metric per instance which is per node. It has multiple
dimensions based on the method and bucket for the latency. For
each node the cardinality observed for this metric was 240