CI/CD: upgrade bot improvements from v0.131.0 PR (#1260)
* Updated the otelcollector upgrade script to also update the Prometheus
package versions in `test/ginkgo-e2e/utils/go.mod` and
`test/ginkgo-e2e/prometheusui/go.mod` to match those used by the
otelcollector and targetallocator. This is necessary for the config
validation tests to have the same version
* Added a check for the targetallocator in addition to the otelcollector
to see if the major.minor golang version in the build pipeline needs to
be updated. Curl the MCR repository for golang images to get the latest
patch version
* Add a new test CR template for the upgrade cluster so that the tests
point to the ginkgo changes in the branch and not in the main branch
* Added the prompts I used and the results from Copilot to update the
script as examples for the future