fix: KSM on Arc, CI/CD retina version, release prod cluster deployment (#1306)
- Move the new optional settings config for KSM outside of the
isArcExtension check so that it is also available for Arc. The install
was failing on our CI/CD arc cluster
- Retina has added pre-release versions to their github that broke our
CI/CD release handling. Get the latest excluding these versions that
don't have the container image version available
- Fix some issues with the release deploying to the CI/CD prod clusters.
We cannot automatically deploy after the image push task because that
waits 24 hours for SDP

Added a task to add dry-run here:
https://dev.azure.com/msazure/InfrastructureInsights/_sprints/taskboard/VMICIAndProm/InfrastructureInsights/Bromine/CY25Q3/Monthly/09%20Sep%20(Sep%2001%20-%20Sep%2030)?workitem=35049643