release: update internal doc + fix usage of new Arc API for release trains (#1339)
- Update internal docs for release that after canary you have to go into
your SAW and manually promote to prod for SDP.
- Managed SDP uses a new release API version for Arc PUT/PATCH calls
that changes the way the versioning is handled. Before, you had to
include all release trains for a version in the call. Now you cannot
re-use them in the call for regions. So the call for stable for
westcentralus where we have a "pipeline" release train already with the
same version is now considered invalid. Fixed this by doing by default:
- Before: `"pipeline","staging","stable"` (had to include all of them if
pipeline and staging already existed for that version in westcentralus)
- Now: `"stable"` (releases stable train for the version in
westcentralus and keeps existing release in pipeline and staging trains)
- Allow to include multiple release trains as a parameter if needed and
parse these correctly for the necessary JSON formatting